requirements.txt
pre-commit
rich
openai
tiktoken
torch
transformers
python-dotenv
click
PyYAML
qiskit==0.43.1
toml
pyparsing
fastcov
genhtml
matplotlib
coverage==7.2.7
pandas==2.0.3
